Ref: Доработка задания: Список фильмов и ShowMore - удаляет HOC
